bugfix: last_create_msg should be u64 (stores a u64 time), not u32

+++ ccp.h
@@ -83,7 +83,7 @@ struct ccp_connection {
     // the index of this array element
     u16 index;
 
-    u32 last_create_msg_sent;
+    u64 last_create_msg_sent;
 
     // struct ccp_primitives is large; as a result, we store it inside ccp_connection to avoid
     // potential limitations in the datapath
